+++ src/template/rendering/htmlAttributeSetter.js
@@ -1,11 +1,12 @@
 
 Class("wipeout.template.rendering.htmlAttributeSetter", function () {
 	
-	function htmlAttributeSetter (name, value, action /* optional */) {
-		this.name = name;
+	var htmlAttributeSetter = wipeout.template.setter.extend(function htmlAttributeSetter (name, value, action /* optional */) {
+		
+		this._super(name, value);
+		
 		this.action = action;
-		this.value = value;
-	}
+	});
 	
 	htmlAttributeSetter.prototype.eventBuild = function () {
 		
@@ -35,23 +36,6 @@ Class("wipeout.template.rendering.htmlAttributeSetter", function () {
 		return output.dispose.bind(output);
     };
 	
-	htmlAttributeSetter.prototype.build = function () {
-		
-		return this._built || (this._built = wipeout.template.context.buildGetter(this.value));
-	};
-	
-	htmlAttributeSetter.prototype.watch = function (renderContext, callback, evaluateImmediately) {
-		if (!this._caching)
-			throw "The watch function can only be called in the context of a cacheAllWatched call. Otherwise the watcher object will be lost, causing memory leaks";
-				
-		var watched = wipeout.utils.htmlBindingTypes.isSimpleBindingProperty(this.value) ?
-			new obsjs.observeTypes.pathObserver(renderContext, this.value) :
-			renderContext.getComputed(this.build());
-		
-		this._caching.push(watched);
-		return watched.onValueChanged(callback, evaluateImmediately);
-	};
-	
 	htmlAttributeSetter.prototype.apply = function (element, renderContext) {
 		var op = [];
 		op.push.apply(op, this.cacheAllWatched((function () {
@@ -63,23 +47,6 @@ Class("wipeout.template.rendering.htmlAttributeSetter", function () {
 		}).bind(this)));
 		
 		return op;
-	};
-	
-	htmlAttributeSetter.prototype.execute = function (renderContext) {
-		return this.build().apply(null, renderContext.asGetterArgs());
-	};
-	
-	htmlAttributeSetter.prototype.cacheAllWatched = function (logic) {
-		if (this._caching)
-			throw "cacheAllWatched cannot be asynchronus or nested.";
-		
-		try {
-			this._caching = [];
-			logic();
-			return this._caching;
-		} finally {
-			delete this._caching;
-		}
 	};
 	
 	return htmlAttributeSetter;
